+++ CHANGELOG.md
@@ -6,12 +6,14 @@ New features:
 
 - #886 The validation bar is pined on the Grid View page
 - #886 Display validation icon after making a validation on the Grid View page
+- #892 The user is able to search pattern matchings jobs by name and template
 
 Resolved issues:
 
 - #886 Fixed the issue with removing selected clusters frequency in the filter
 - #886 Fixed the issue with a chronological order of audio event detection jobs
 - #886 Unselect selected boxes after making the validation on the Grid View page
+- #892 Fixed the issue related with searching pattern matchings jobs by name and template
 
 ## v3.0.47 - February 17, 2022